vr-backend/rembg-api/requirements.txt
fastapi
uvicorn
pillow
numpy
rembg
torch
torchvision
